Creating a simple bottom nav-bar along with leave squad pop up menu
