+++ squadlight/lib/main.dart
@@ -5,118 +5,90 @@ import 'package:flutter_map/flutter_map.dart';
 import 'package:geolocator/geolocator.dart';
 import 'package:latlong2/latlong.dart';
 import 'socket.dart';
+import 'pages/home/home.dart';
+import 'pages/map/map.dart';
+import 'pages/chat/chat.dart';
 
 void main() {
   runApp(
     const MaterialApp(
-      home: MapPage(),
+      home: MainPage(),
     ),
   );
 }
 
-class MapPage extends StatefulWidget {
-  const MapPage({Key? key}) : super(key: key);
+class MainPage extends StatefulWidget {
+  const MainPage({Key? key}) : super(key: key);
 
   @override
-  State<MapPage> createState() => _MapPageState();
+  State<MainPage> createState() => _MainPageState();
 }
 
-class _MapPageState extends State<MapPage> {
-  LatLng userLoc = LatLng(53.472164, -2.238193);
+class _MainPageState extends State<MainPage> {
+// Navigation - Current Selected Item
+  int _selectedIndex = 0;
 
-  /// Determine the current position of the device.
-  ///
-  /// When the location services are not enabled or permissions
-  /// are denied the `Future` will return an error.
-  Future<Position> _determinePosition() async {
-    bool serviceEnabled;
-    LocationPermission permission;
-
-    // Test if location services are enabled.
-    serviceEnabled = await Geolocator.isLocationServiceEnabled();
-    if (!serviceEnabled) {
-      // Location services are not enabled don't continue
-      // accessing the position and request users of the
-      // App to enable the location services.
-      return Future.error('Location services are disabled.');
-    }
-
-    permission = await Geolocator.checkPermission();
-    if (permission == LocationPermission.denied) {
-      permission = await Geolocator.requestPermission();
-      if (permission == LocationPermission.denied) {
-        // Permissions are denied, next time you could try
-        // requesting permissions again (this is also where
-        // Android's shouldShowRequestPermissionRationale
-        // returned true. According to Android guidelines
-        // your App should show an explanatory UI now.
-        return Future.error('Location permissions are denied');
-      }
-    }
-
-    if (permission == LocationPermission.deniedForever) {
-      // Permissions are denied forever, handle appropriately.
-      return Future.error(
-          'Location permissions are permanently denied, we cannot request permissions.');
-    }
-
-    // When we reach here, permissions are granted and we can
-    // continue accessing the position of the device.
-    return await Geolocator.getCurrentPosition();
-  }
-
-  _getCurrentLocation() {
-    Geolocator.getCurrentPosition(
-            desiredAccuracy: LocationAccuracy.best,
-            forceAndroidLocationManager: true)
-        .then((Position position) {
-      setState(() {
-        _currentPosition = position;
-        userLoc = LatLng(_currentPosition.latitude, _currentPosition.longitude);
-      });
-    }).catchError((e) {
-      print(e);
+// Navigation - On Tap Switch Selection
+  void _onItemTapped(int index) {
+    setState(() {
+      _selectedIndex = index;
     });
   }
 
-  late Position _currentPosition;
+// Navigation - Pages
+  List<Widget> _pages = <Widget>[
+    // HomePage(),
+    ChatPage(),
+    MapPage(),
+  ];
+
+  var appBarHeight = AppBar().preferredSize.height;
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      floatingActionButton: FloatingActionButton(
-          onPressed: () {
-            _getCurrentLocation();
-          },
-          child: const Text('Get User Location')),
-      body: Center(
-        child: Column(
-          children: [
-            Text("Location Here"),
-            Flexible(
-              child: FlutterMap(
-                options: MapOptions(
-                  center: LatLng(53.472164, -2.238193),
-                  zoom: 8,
-                ),
-                layers: [
-                  TileLayerOptions(
-                      urlTemplate:
-                          "https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png",
-                      subdomains: ['a', 'b', 'c']),
-                  MarkerLayerOptions(
-                    markers: [
-                      Marker(
-                        point: LatLng(userLoc.latitude, userLoc.longitude),
-                        builder: (ctx) => const Icon(Icons.pin_drop),
-                      ),
-                    ],
-                  ),
-                ],
-              ),
-            ),
-          ],
-        ),
+      appBar: AppBar(
+        title: Text("SquadLight"),
+        backgroundColor: Colors.grey[850],
+        actions: [
+          PopupMenuButton(
+            offset: Offset(0.0, appBarHeight),
+            itemBuilder: (context) => [
+              PopupMenuItem(
+                child: Text('Leave Squad'),
+                onTap: () {
+                  print("home screen");
+                },
+              )
+            ],
+          )
+        ],
+      ),
+
+      body: IndexedStack(
+        index: _selectedIndex,
+        children: _pages,
+      ),
+      bottomNavigationBar: BottomNavigationBar(
+        backgroundColor: Colors.grey[850],
+        unselectedItemColor: Colors.grey,
+        selectedFontSize: 16,
+        selectedIconTheme: IconThemeData(color: Colors.amberAccent, size: 40),
+        selectedItemColor: Colors.amberAccent,
+        selectedLabelStyle: TextStyle(fontWeight: FontWeight.bold),
+        iconSize: 30,
+        items: const <BottomNavigationBarItem>[
+          BottomNavigationBarItem(
+            icon: Icon(Icons.forum),
+            label: "Chat",
+          ),
+          BottomNavigationBarItem(
+            icon: Icon(Icons.map),
+            label: "Map",
+          ),
+        ],
+        currentIndex: _selectedIndex,
+        onTap: _onItemTapped,
       ),
     );
   }
