+++ squadlight/lib/pages/home/home.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class HomePage extends StatelessWidget {
+  const HomePage({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Center(
+        child: ElevatedButton(
+          onPressed: () {},
+          child: Text('Welcome To The Home Page'),
+        ),
+      ),
+    );
+  }
+}
