--- squadlight/pages/chat/chat.dart
@@ -1,15 +1,0 @@
-import 'package:flutter/material.dart';
-
-class ChatPage extends StatefulWidget {
-  const ChatPage({Key? key}) : super(key: key);
-
-  @override
-  State<ChatPage> createState() => _ChatPageState();
-}
-
-class _ChatPageState extends State<ChatPage> {
-  @override
-  Widget build(BuildContext context) {
-    return Container();
-  }
-}
