+++ squadlight/test/widget_test.dart
@@ -1,3 +1,5 @@
+
+
 // This is a basic Flutter widget test.
 //
 // To perform an interaction with a widget in your test, use the WidgetTester
@@ -9,6 +11,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'package:squadlight/main.dart';
+import '../lib/pages/map/map.dart';
 
 void main() {
   testWidgets('Counter increments smoke test', (WidgetTester tester) async {
